BIG VUFORIA STUFF maybe?
So I searched a video to work on vuforia and found this: https://www.youtube.com/watch?v=AxKrJEtfuaI. But yeah, so I followed along and typed my code as he went through and used his repository for reference. I readjsusted variables and code to hopefully apply to our personal robot (of course will be updated along with the building of the robot itself). and I added as much commenting as I could to try to explain what I understood from the video. If this works, we'll have taken a big step into understanding autonomous. If not, eh. I was bored anyways.